# fix(#911): [Salvage PR3/P1] Domain Value Objects