--- a/core/src/lib.rs
+++ b/core/src/lib.rs
@@ -14,6 +14,7 @@ pub mod config;
 pub mod consumables;
 pub mod deck;
 pub mod display;
+pub mod domain;
 pub mod error;
 pub mod game;
 pub mod generator;