## Summary

Implements domain value objects salvaged from closed PR #776 as part of the Domain Foundation salvage operation. This is P1 priority issue #911, implementing fundamental domain building blocks.

## Changes

### Value Objects Implemented (~350 lines total)

1. **SessionId Value Object** (~75 lines)
   - Strongly-typed wrapper for session IDs
   - UUID-based with validation
   - Immutable with proper trait implementations

2. **ValidationResult Value Object** (~50 lines)
   - Encapsulates validation outcomes
   - Success/failure states with detailed error information
   - Clean error handling pattern

3. **Money Value Object** (~75 lines)
   - Domain-specific money handling
   - Prevents negative values at type level
   - Safe arithmetic operations with overflow protection
   - Can't go negative through subtraction

4. **Score Value Object** (~75 lines)
   - Encapsulates scoring logic
   - Prevents invalid score states
   - Supports chips/mult calculations
   - Formatted display with thousands separators

5. **Module Organization** (~75 lines)
   - Created `core/src/domain/value_objects/` structure
   - Common traits for value objects
   - Proper module exports and re-exports

## Technical Details

- **Zero-cost abstractions**: Newtype pattern with no runtime overhead
- **Immutable by design**: All value objects are immutable
- **Validation at construction**: Invariants enforced at creation time
- **Comprehensive tests**: 25 tests covering all value objects
- **Type safety**: Strong typing prevents invalid states

## Testing

```bash
cargo test domain --lib
# 25 tests pass

cargo clippy -- -D warnings
# Zero warnings

cargo build
# Compiles successfully
```

## Success Criteria Met

✅ Immutable value objects with validation
✅ All code compiles: `cargo build`
✅ All tests pass: `cargo test`
✅ Zero clippy warnings: `cargo clippy -- -D warnings`
✅ Under 350 lines total
✅ Kernel-quality standards applied

## Related Issues

- Resolves #911
- Part of Domain Foundation salvage operation (master issue #906)
- Depends on #909 (merged)
- Works in parallel with #910 (Domain Infrastructure)

## Next Steps

After merge, the next P1 issues are:
- #912: [Salvage PR3/P1] Domain Services
- #913: [Salvage PR3/P1] Domain Entities

---

*Implementation follows kernel-quality standards with zero tolerance for sloppiness.*